Add dummy function for R CMD Check
Dummy function just to make one explicit usage of R.utils (dependency)
to please R CMD check test and avoid it complaining that R.utils is
listed in the Imports section of DESCRIPTION but it is not being used
anywhere.
R.utils is a "cross dependency" from piggyback.
Got this trick from:
https://community.rstudio.com/t/new-r-cmd-check-note-in-r-4-2-0-for-imports-field/143153/2