Do not run slow examples.
CRAN has a limit of 5 seconds to run the examples. Since we rely on
downloading large files, it usually take more than 5 seconds as a
default, failing the "Checking examples" check from CRAN. So, the
proposed approach here is to "Skip"  running these examples.

Due to the fact that we indeed have tests covering these functions,
skipping this "doc examples" execution won't impact in the project
coverage.