Undo force_download
By default piggyback:pb_download check the local and remote file hash to
ensure it will only download the file if it is a new file (different
hash). So we don't need this "foce_download" flag all around.

The only "requirement" here is that we need to ensure the same path/file
will be used across all executions.